fix(settings): surface Compute Host removal failures
Use the Compute list row action as the single removal entry point. Keep eligibility checks, password-deletion confirmation, and deletion errors in that X-triggered dialog so failures remain visible and retryable.
